requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=4.9.0

# OpenAI
openai>=1.0.0

# LangChain and LangGraph
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5

# Vector database
faiss-cpu>=1.7.4

# Embeddings and image processing
sentence-transformers>=2.2.0
Pillow>=10.0.0
numpy>=1.24.0

# Image comparison
scikit-image>=0.21.0

# Token counting
tiktoken>=0.5.0

# Type hints
typing-extensions>=4.8.0

